--- src/User.ts
@@ -1,6 +1,0 @@
-export interface User {
-  id: number;
-  name: string;
-  email: string;
-  birthDate: string;
-}
